requirements.txt
requests
astroid==2.3.3
certifi==2019.11.28
Django==2.2.5
django-chartjs==1.5.0
django-model-utils==3.2.0
django-pandas==0.6.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wrapt==1.11.2
django-bootstrap-select==0.1.3
django-appconf==1.0.3
django-select2==7.2.0
statsmodels
libsass 
django-compressor 
django-sass-processor
plotly
